Make hpos-admin a single-file python command
