+++ modules/services/hpos-admin.nix
@@ -23,7 +23,7 @@ in
     environment.systemPackages = with pkgs; [
       hpos-state-gen-cli # for testing
     ];
-    systemd.services.holo-admin = {
+    systemd.services.hpos-admin = {
       environment.HPOS_STATE_PATH = "/tmp/hpos-state.json";
       after = [ "network.target" ];
       wantedBy = [ "multi-user.target" ];
