+++ overlays/holo-nixpkgs/hpos-admin/hpos-admin.py
@@ -4,8 +4,11 @@ from hashlib import sha512
 import json
 import logging
 import os
-from gevent import subprocess
 from datetime import datetime
+
+from gevent import Greenlet, sleep, subprocess
+from gevent.event import Event
+from gevent.subprocess import Popen, PIPE
 
 log = logging.getLogger(__name__)
 
@@ -32,7 +35,68 @@ def get_config():
 
 # APIs that require a nixos-rebuild after mutating HPOS state may trigger it using rebuild.go().
 # This is a gevent-compatible Greenlet scheduler, that ensures only one invocation runs at a time.
-from runner import Runner
+
+class Runner( Greenlet ):
+    """Performs a command (which is/returns a list) on demand from multiple parties, but only once at a
+    time.  In other words, when we detect that we're supposed to start the command, we'll loop
+    around and wait 'til anyone else calls (or has already called) self.go().
+
+    """
+    def __init__( self, command, *args, **kwds ):
+        self._event		= Event()
+        self._command		= command # callable or list
+        self._done		= False
+        self.trigger		= 0
+        self.counter		= 0
+        super().__init__( *args, **kwds )
+        log.info(f"{self} Starting")
+
+    def __str__( self ):
+        return f"Runner({self.command})"
+
+    def __del__( self ):
+        self.done		= True
+        self.join()
+
+    @property
+    def command( self ):
+        if hasattr(self._command, '__call__'):
+            return self._command(self)
+        return self._command
+
+    @property
+    def done( self ):
+        return self._done
+    @done.setter
+    def done( self, value ):
+        self._done		= bool( value )
+        if self._done:
+            self.go()
+
+    def go( self ):
+        self.trigger	       += 1
+        self._event.set()
+
+    def _run( self ):
+        """Await Event.set() via self.go(), and loop 'til self.done is set"""
+        log.info(f"{self}: Ready to go")
+        try:
+            while self._event.wait() and not self.done:
+                self.counter   += 1
+                log.info(f"{self}: Run {self.counter}, w/ {self.trigger} triggers")
+                self._event.clear()
+                # Calls to self.go() *after* this will point will cause a further execution of command!
+                try:
+                    status 	= Popen( self.command )
+                    status.wait()
+                    (log.warning if status.returncode else log.info)(
+                        f"{self}: Run {self.counter} Exit: {status.returncode}" )
+                except Exception as exc:
+                    log.warning(
+                        f"{self}: Run {self.counter} Exception: {exc}" )
+        finally:
+            log.info(f"{self}: Finished after {self.counter} runs, {self.trigger} triggers")
+
 rebuild = Runner( ['sudo', 'nixos-rebuild', '--upgrade', 'switch'] )
 rebuild.start()
 
@@ -51,9 +115,9 @@ def put_config():
 
 
 def zerotier_info():
-    proc = subprocess.Popen(
+    proc = Popen(
         [ 'sudo', 'zerotier-cli', '-j', 'info' ],
-        stdout=subprocess.PIPE, stderr=subprocess.PIPE
+        stdout=PIPE, stderr=PIPE
     )
     stdout, stderr = proc.communicate()
     assert not proc.returncode, \
